perf(vm): implement tail call optimization for deep recursion
【动机】
当前函数调用在尾部位置（如 return f()）仍分配新的调用帧，导致递归深度受限，深层递归可能栈溢出。同时重复的帧分配也带来性能开销。

【原理】
在编译器中识别尾部位置（return 语句中直接返回另一个函数调用），生成新的尾调用指令（如 OP_TAIL_CALL）。虚拟机执行尾调用时复用当前调用帧，替换函数和参数，不增加栈深度，从而支持任意深度的尾递归并减少帧分配开销。

+++ src/vm/VM.cpp
@@ -2143,7 +2143,7 @@ namespace jc {
                                     result = inst->fields->elements[it->second].second;
                                     found = true;
                                     ic.cachedClass = inst->classDef;
-                                    ic.cachedFieldIndex = it->second;
+                                    ic.cachedFieldIndex = static_cast<int>(it->second);
                                     ic.cachedMethod = nullptr;
                                 }
                             }
@@ -2323,7 +2323,7 @@ namespace jc {
                                     result = inst->fields->elements[it->second].second;
                                     found = true;
                                     ic.cachedClass = inst->classDef;
-                                    ic.cachedFieldIndex = it->second;
+                                    ic.cachedFieldIndex = static_cast<int>(it->second);
                                 }
                             }
                             if (!found) {
@@ -2400,10 +2400,10 @@ namespace jc {
                             if (it != inst->fields->keyMap.end()) {
                                 inst->fields->elements[it->second].second = val;
                                 ic.cachedClass = inst->classDef;
-                                ic.cachedFieldIndex = it->second;
+                                ic.cachedFieldIndex = static_cast<int>(it->second);
                             } else {
                                 ic.cachedClass = inst->classDef;
-                                ic.cachedFieldIndex = inst->fields->elements.size();
+                                ic.cachedFieldIndex = static_cast<int>(inst->fields->elements.size());
                                 inst->fields->keyMap[key] = inst->fields->elements.size();
                                 inst->fields->elements.push_back({key, val});
                             }
